+++ python/tests/conftest.py
@@ -1,7 +1,10 @@
-import pytest
-import tempfile
+import json
 import shutil
-import os
+import tempfile
+
+import pytest
+
+from .helpers import REFERENCE_GENOMES_FILE
 
 
 @pytest.fixture
@@ -17,3 +20,19 @@ def empty_database():
     """Create an empty database instance."""
     from rhydb import Database
     return Database()
+
+
+@pytest.fixture
+def reference_genomes():
+    """Load reference genomes from test data."""
+    with open(REFERENCE_GENOMES_FILE, 'r') as f:
+        return json.load(f)
+
+
+@pytest.fixture
+def main_reference_sequence(reference_genomes):
+    """Get the main nucleotide reference sequence."""
+    for seq in reference_genomes['nucleotideSequences']:
+        if seq['name'] == 'main':
+            return seq['sequence']
+    raise ValueError("No 'main' sequence found in reference genomes")
