+++ python/tests/test_database.py
@@ -1,82 +1,6 @@
+"""Tests for module import, database construction and loading."""
+
 import pytest
-import os
-import json
-import datetime
-import tempfile
-import shutil
-import pyroaring
-import pyarrow as pa
-
-
-# Path to test data
-TEST_DATA_DIR = os.path.join(os.path.dirname(__file__), '..', '..', 'testBaseData', 'exampleDataset')
-REFERENCE_GENOMES_FILE = os.path.join(TEST_DATA_DIR, 'reference_genomes.json')
-INPUT_FILE = os.path.join(TEST_DATA_DIR, 'input_file.ndjson')
-
-# A serialized database (checked into the repo) whose schema has scalar value columns
-# (age:int, qc_value:float, date:date, test_boolean_column:bool). Scalar columns can now also be
-# created directly through create_table; loading this state additionally exercises the on-disk
-# load path for such columns.
-SERIALIZED_STATE_DIR = os.path.join(
-    os.path.dirname(__file__), '..', '..', 'testBaseData', 'siloSerializedState'
-)
-
-
-@pytest.fixture
-def reference_genomes():
-    """Load reference genomes from test data."""
-    with open(REFERENCE_GENOMES_FILE, 'r') as f:
-        return json.load(f)
-
-
-@pytest.fixture
-def main_reference_sequence(reference_genomes):
-    """Get the main nucleotide reference sequence."""
-    for seq in reference_genomes['nucleotideSequences']:
-        if seq['name'] == 'main':
-            return seq['sequence']
-    raise ValueError("No 'main' sequence found in reference genomes")
-
-
-def _register_reference(db, name, reference):
-    """Register a sequence column's reference in the built-in ``reference_genomes`` table.
-
-    The generic ``create_table`` interface resolves sequence-column references from the built-in
-    ``reference_genomes`` table rather than inline, so the reference must be present before the
-    sequence table is created. Uses the ``register_reference`` short-hand.
-    """
-    db.register_reference(name, reference)
-
-
-def create_nucleotide_sequence_table(
-    db, table_name, primary_key_name, sequence_name, reference_sequence, extra_columns=None
-):
-    """Create a nucleotide sequence table through the generic ``create_table`` interface.
-
-    Replaces the removed ``create_nucleotide_sequence_table`` short-hand: the primary key is the
-    first column, followed by any extra string columns and the nucleotide sequence column, whose
-    reference is registered in the built-in ``reference_genomes`` table.
-    """
-    _register_reference(db, sequence_name, reference_sequence)
-    columns = [{"name": primary_key_name, "type": "string"}]
-    columns += [{"name": col, "type": "string"} for col in (extra_columns or [])]
-    columns.append({"name": sequence_name, "type": "nucleotide_sequence"})
-    db.create_table(table_name, columns)
-
-
-def create_gene_table(
-    db, table_name, primary_key_name, gene_name, reference_sequence, extra_columns=None
-):
-    """Create an amino acid sequence table through the generic ``create_table`` interface.
-
-    Replaces the removed ``create_gene_table`` short-hand (see
-    :func:`create_nucleotide_sequence_table`), using an amino acid sequence column.
-    """
-    _register_reference(db, gene_name, reference_sequence)
-    columns = [{"name": primary_key_name, "type": "string"}]
-    columns += [{"name": col, "type": "string"} for col in (extra_columns or [])]
-    columns.append({"name": gene_name, "type": "amino_acid_sequence"})
-    db.create_table(table_name, columns)
 
 
 class TestDatabaseImport:
@@ -132,285 +56,6 @@ class TestDatabaseCreation:
             assert db is not None
 
 
-class TestCreateNucleotideSequenceTable:
-    """Test creating nucleotide sequence tables."""
-
-    def test_create_table_with_simple_reference(self, empty_database):
-        """Test creating a table with a simple reference sequence."""
-        # Table names must be lowercase
-        create_nucleotide_sequence_table(empty_database,
-            table_name="testtable",
-            primary_key_name="id",
-            sequence_name="main",
-            reference_sequence="ACGTACGTACGT"
-        )
-        # If no exception, table was created
-
-    def test_create_table_with_real_reference(self, empty_database, main_reference_sequence):
-        """Test creating a table with the real SARS-CoV-2 reference sequence."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="gisaidepisl",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        # If no exception, table was created
-
-    def test_get_reference_sequence_after_create(self, empty_database, main_reference_sequence):
-        """Test that we can retrieve the reference sequence after creating a table."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-
-        retrieved = empty_database.get_nucleotide_reference_sequence("sequences", "main")
-        assert retrieved == main_reference_sequence
-
-    def test_get_gene_reference_sequence_after_create(self, empty_database):
-        """Test that we can retrieve the reference sequence after creating a table."""
-        create_gene_table(empty_database,
-            table_name="sequences",
-            primary_key_name="key",
-            gene_name="main",
-            reference_sequence="ABCD"
-        )
-
-        retrieved = empty_database.get_amino_acid_reference_sequence("sequences", "main")
-        assert retrieved == "ABCD"
-
-
-class TestCreateGeneTable:
-    """Test creating gene tables."""
-
-    def test_create_gene_table(self, empty_database):
-        """Test creating a gene table."""
-        # Table names must be lowercase
-        create_gene_table(empty_database,
-            table_name="genes",
-            primary_key_name="id",
-            gene_name="S",
-            reference_sequence="MFVFLVLLPLVSSQCVNLTTRTQLPPAYTNSFTRGVYYPDKVFRSSVLHSTQDLFLPFFSNVTWFHAI*"
-        )
-        # If no exception, table was created
-
-
-class TestAppendData:
-    """Test appending data to tables."""
-
-    def test_append_data_from_real_file(self, empty_database, main_reference_sequence):
-        """Test appending data from the real test data file."""
-        # First create the table
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-
-        # Then append data
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-        # If no exception, data was appended
-
-
-class TestGetFilteredBitmap:
-    """Test getting filtered bitmaps."""
-
-    def test_get_filtered_bitmap_true_filter(self, empty_database, main_reference_sequence):
-        """Test getting a bitmap with True filter (returns all rows)."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        # True filter should return all rows
-        bitmap = empty_database.get_filtered_bitmap("sequences", 'true')
-        assert isinstance(bitmap, pyroaring.BitMap)
-        assert len(bitmap) > 0  # Should have at least one row from test data
-
-    def test_get_filtered_bitmap_returns_bitmap(self, empty_database, main_reference_sequence):
-        """Test that get_filtered_bitmap returns a pyroaring.BitMap."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        bitmap = empty_database.get_filtered_bitmap("sequences", 'true')
-        assert isinstance(bitmap, pyroaring.BitMap)
-        # Can iterate over bitmap to get indices
-        indices = list(bitmap)
-        for idx in indices:
-            assert isinstance(idx, int)
-
-    def test_get_filtered_bitmap_with_none_filter(self, empty_database, main_reference_sequence):
-        """Test that None filter defaults to True filter."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        bitmap = empty_database.get_filtered_bitmap("sequences", None)
-        assert isinstance(bitmap, pyroaring.BitMap)
-
-    def test_get_filtered_bitmap_with_empty_filter(self, empty_database, main_reference_sequence):
-        """Test that empty string filter defaults to True filter."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        bitmap = empty_database.get_filtered_bitmap("sequences", "")
-        assert isinstance(bitmap, pyroaring.BitMap)
-        assert len(bitmap) > 0
-
-    def test_get_filtered_bitmap_supports_set_operations(self, empty_database, main_reference_sequence):
-        """Test that returned bitmap supports set operations."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        bitmap = empty_database.get_filtered_bitmap("sequences", 'true')
-
-        # Test union, intersection operations
-        other_bitmap = pyroaring.BitMap([0, 1, 2])
-        union = bitmap | other_bitmap
-        intersection = bitmap & other_bitmap
-
-        assert isinstance(union, pyroaring.BitMap)
-        assert isinstance(intersection, pyroaring.BitMap)
-
-class TestSaveAndLoadCheckpoint:
-    """Test saving and loading database checkpoints."""
-
-    def test_save_checkpoint(self, empty_database, main_reference_sequence, temp_dir):
-        """Test saving a database checkpoint."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        save_path = os.path.join(temp_dir, "checkpoint")
-        empty_database.save_checkpoint(save_path)
-
-        # Check that something was saved
-        assert os.path.exists(save_path)
-
-    def test_load_from_checkpoint(self, empty_database, main_reference_sequence, temp_dir):
-        """Test loading a database from a checkpoint."""
-        from rhydb import Database
-
-        # Create and save
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        save_path = os.path.join(temp_dir, "checkpoint")
-        empty_database.save_checkpoint(save_path)
-
-        # Load from checkpoint
-        loaded_db = Database(save_path)
-        assert loaded_db is not None
-
-        # Verify data is preserved
-        ref_seq = loaded_db.get_nucleotide_reference_sequence("sequences", "main")
-        assert ref_seq == main_reference_sequence
-
-    def test_checkpoint_preserves_data(self, empty_database, main_reference_sequence, temp_dir):
-        """Test that checkpoint preserves all data correctly."""
-        from rhydb import Database
-
-        # Create, add data, and save
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        # Get data before save
-        bitmap_before = empty_database.get_filtered_bitmap("sequences", 'true')
-
-        # Save and reload
-        save_path = os.path.join(temp_dir, "checkpoint")
-        empty_database.save_checkpoint(save_path)
-
-        loaded_db = Database(save_path)
-
-        # Compare with loaded data
-        bitmap_after = loaded_db.get_filtered_bitmap("sequences", 'true')
-
-        assert bitmap_before == bitmap_after
-
-
-class TestDatabaseValidation:
-    """Test input validation for database methods."""
-
-    def test_append_data_empty_table_name(self, empty_database):
-        """Test that empty table name raises ValueError."""
-        with pytest.raises(ValueError, match="table_name cannot be empty"):
-            empty_database.append_data_from_file("", "/some/file.ndjson")
-
-    def test_append_data_empty_file_name(self, empty_database):
-        """Test that empty file name raises ValueError."""
-        with pytest.raises(ValueError, match="file_name cannot be empty"):
-            empty_database.append_data_from_file("test_table", "")
-
-    def test_append_data_nonexistent_file(self, empty_database):
-        """Test that nonexistent file raises FileNotFoundError."""
-        with pytest.raises(FileNotFoundError):
-            empty_database.append_data_from_file("test_table", "/nonexistent/file.ndjson")
-
-    def test_print_all_data_empty_table_name(self, empty_database):
-        """Test that empty table name raises ValueError."""
-        with pytest.raises(ValueError, match="table_name cannot be empty"):
-            empty_database.print_all_data("")
-
-    def test_get_nucleotide_reference_empty_table_name(self, empty_database):
-        """Test that empty table name raises ValueError."""
-        with pytest.raises(ValueError, match="table_name cannot be empty"):
-            empty_database.get_nucleotide_reference_sequence("", "main")
-
-    def test_get_nucleotide_reference_empty_sequence_name(self, empty_database):
-        """Test that empty sequence name raises ValueError."""
-        with pytest.raises(ValueError, match="sequence_name cannot be empty"):
-            empty_database.get_nucleotide_reference_sequence("test", "")
-
-    def test_get_filtered_bitmap_empty_table_name(self, empty_database):
-        """Test that empty table name raises ValueError."""
-        with pytest.raises(ValueError, match="table_name cannot be empty"):
-            empty_database.get_filtered_bitmap("", "filter")
-
-    def test_save_checkpoint_empty_directory(self, empty_database):
-        """Test that empty directory raises ValueError."""
-        with pytest.raises(ValueError, match="save_directory cannot be empty"):
-            empty_database.save_checkpoint("")
-
-
 class TestDatabaseLoad:
     """Test database loading from saved state."""
 
@@ -419,684 +64,3 @@ class TestDatabaseLoad:
         from rhydb import Database
         with pytest.raises(FileNotFoundError):
             Database("/nonexistent/path/to/database")
-
-
-class TestPrintAllData:
-    """Test the print_all_data method."""
-
-    def test_print_all_data(self, empty_database, main_reference_sequence, capsys):
-        """Test that print_all_data outputs something."""
-        # Note: printAllData expects sequence_name="sequence" (hardcoded in C++)
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        # This should print to stdout
-        empty_database.print_all_data("sequences")
-
-        # Check that something was printed
-        captured = capsys.readouterr()
-        # The output goes to C++ stdout, so we may not capture it in Python
-        # Just verify no exception was raised
-
-
-class TestExtraColumns:
-    """Test creating tables with extra string columns."""
-
-    def test_create_table_with_extra_columns(self):
-        """Test creating a table with extra string columns."""
-        from rhydb import Database
-
-        db = Database()
-        create_nucleotide_sequence_table(db,
-            table_name="test",
-            primary_key_name="id",
-            sequence_name="seq",
-            reference_sequence="ACGT",
-            extra_columns=["country", "date", "lineage"]
-        )
-        # If no exception, table was created with extra columns
-
-    def test_extra_columns_accept_data(self):
-        """Test that extra columns can store and retrieve data."""
-        from rhydb import Database
-
-        db = Database()
-        create_nucleotide_sequence_table(db,
-            table_name="test",
-            primary_key_name="id",
-            sequence_name="seq",
-            reference_sequence="ACGT",
-            extra_columns=["country", "lineage"]
-        )
-        db.append_data_from_string("test", '{"id": "s1", "seq": {"sequence": "AAAA", "insertions": []}, "country": "USA", "lineage": "BA.1"}')
-        db.append_data_from_string("test", '{"id": "s2", "seq": {"sequence": "CCCC", "insertions": []}, "country": "UK", "lineage": "BA.2"}')
-
-        query = 'test'
-        result = db.query(query)
-
-        assert "country" in result.column_names
-        assert "lineage" in result.column_names
-        data = result.to_pydict()
-        assert set(data["country"]) == {"USA", "UK"}
-        assert set(data["lineage"]) == {"BA.1", "BA.2"}
-
-    def test_extra_columns_default_empty(self):
-        """Test that extra_columns defaults to empty (backward compatible)."""
-        from rhydb import Database
-
-        db = Database()
-        # Call without extra_columns - should work as before
-        create_nucleotide_sequence_table(db,
-            table_name="test",
-            primary_key_name="id",
-            sequence_name="seq",
-            reference_sequence="ACGT"
-        )
-        db.append_data_from_string("test", '{"id": "s1", "seq": {"sequence": "AAAA", "insertions": []}}')
-
-        query = 'test'
-        result = db.query(query)
-        assert result.num_rows == 1
-
-    def test_extra_columns_with_none(self):
-        """Test that extra_columns=None works."""
-        from rhydb import Database
-
-        db = Database()
-        create_nucleotide_sequence_table(db,
-            table_name="test",
-            primary_key_name="id",
-            sequence_name="seq",
-            reference_sequence="ACGT",
-            extra_columns=None
-        )
-        db.append_data_from_string("test", '{"id": "s1", "seq": {"sequence": "AAAA", "insertions": []}}')
-
-        query = 'test'
-        result = db.query(query)
-        assert result.num_rows == 1
-
-    def test_extra_columns_invalid_type_raises(self):
-        """Test that a non-string extra column is rejected by create_table's validation."""
-        from rhydb import Database
-
-        db = Database()
-        with pytest.raises(ValueError, match="non-empty string 'name'"):
-            create_nucleotide_sequence_table(db,
-                table_name="test",
-                primary_key_name="id",
-                sequence_name="seq",
-                reference_sequence="ACGT",
-                extra_columns=["valid", 123]  # 123 is not a string
-            )
-
-    def test_gene_table_with_extra_columns(self):
-        """Test creating a gene table with extra columns."""
-        from rhydb import Database
-
-        db = Database()
-        create_gene_table(db,
-            table_name="genes",
-            primary_key_name="id",
-            gene_name="spike",
-            reference_sequence="MFVFLVLLPLVSSQCVNLTTRTQLPPAYTNSFTRGVYYPDKVFRSSVLHSTQDLFLPFFSNVTWFHAI*",
-            extra_columns=["variant", "source"]
-        )
-        # If no exception, table was created
-
-
-class TestCreateTable:
-    """Test the generic create_table method that accepts columns of any supported type."""
-
-    @staticmethod
-    def _populate_references(db, entries):
-        """Populate the built-in `reference_genomes` table create_table reads references from.
-
-        `entries` maps a sequence column name to its reference string.
-        """
-        for name, reference in entries.items():
-            db.register_reference(name, reference)
-
-    def test_create_table_with_all_column_types(self):
-        """A table can be created with every supported column type and populated."""
-        from rhydb import Database
-
-        db = Database()
-        self._populate_references(db, {"seq": "ACGT", "gene": "MFV"})
-        db.create_table(
-            table_name="samples",
-            columns=[
-                {"name": "id", "type": "string"},  # first column -> primary key
-                {"name": "age", "type": "int"},
-                {"name": "qc", "type": "float"},
-                {"name": "collected", "type": "date"},
-                {"name": "passed", "type": "bool"},
-                {"name": "country", "type": "string"},
-                {"name": "lineage", "type": "indexed_string"},
-                {"name": "seq", "type": "nucleotide_sequence"},
-                {"name": "gene", "type": "amino_acid_sequence"},
-            ],
-        )
-        db.append_data_from_string(
-            "samples",
-            json.dumps(
-                {
-                    "id": "s1",
-                    "age": 42,
-                    "qc": 0.5,
-                    "collected": "2021-03-15",
-                    "passed": True,
-                    "country": "Switzerland",
-                    "lineage": "B.1",
-                    "seq": {"sequence": "ACGT", "insertions": []},
-                    "gene": {"sequence": "MFV", "insertions": []},
-                }
-            ),
-        )
-
-        result = db.query("samples.project({id, age, qc, country, lineage})")
-        data = result.to_pydict()
-        assert data["id"] == ["s1"]
-        assert data["age"] == [42]
-        assert data["qc"] == [0.5]
-        assert data["country"] == ["Switzerland"]
-        assert data["lineage"] == ["B.1"]
-        # The sequence column's reference was resolved from the `reference_genomes` table by column name.
-        assert db.get_nucleotide_reference_sequence("samples", "seq") == "ACGT"
-        assert db.get_amino_acid_reference_sequence("samples", "gene") == "MFV"
-
-    def test_create_table_reference_looked_up_by_column_name(self):
-        """A sequence column's reference is taken from the matching `reference_genomes` entry."""
-        from rhydb import Database
-
-        db = Database()
-        self._populate_references(db, {"main": "ACGTACGT"})
-        db.create_table(
-            "sequences",
-            [{"name": "id", "type": "string"}, {"name": "main", "type": "nucleotide_sequence"}],
-        )
-
-        assert db.get_nucleotide_reference_sequence("sequences", "main") == "ACGTACGT"
-
-    def test_create_table_unregistered_reference_raises(self, empty_database):
-        """Creating a sequence column with nothing registered in `reference_genomes` fails."""
-        with pytest.raises(RuntimeError, match="no entry named 'seq'"):
-            empty_database.create_table(
-                "samples",
-                [{"name": "id", "type": "string"}, {"name": "seq", "type": "nucleotide_sequence"}],
-            )
-
-    def test_create_table_missing_reference_entry_raises(self):
-        """Creating a sequence column with no matching `reference_genomes` entry fails."""
-        from rhydb import Database
-
-        db = Database()
-        self._populate_references(db, {"other": "ACGT"})
-        with pytest.raises(RuntimeError, match="no entry named 'seq'"):
-            db.create_table(
-                "samples",
-                [{"name": "id", "type": "string"}, {"name": "seq", "type": "nucleotide_sequence"}],
-            )
-
-    def test_create_table_scalar_columns_are_updatable(self):
-        """Scalar columns created via create_table support update_column end to end."""
-        from rhydb import Database
-
-        db = Database()
-        db.create_table(
-            table_name="samples",
-            columns=[{"name": "id", "type": "string"}, {"name": "age", "type": "int"}],
-        )
-        db.append_data_from_string("samples", '{"id": "s1", "age": 1}')
-        db.append_data_from_string("samples", '{"id": "s2", "age": 2}')
-
-        db.update_column("samples", "age", "99", "age = 1")
-
-        assert len(db.get_filtered_bitmap("samples", "age = 99")) == 1
-        assert len(db.get_filtered_bitmap("samples", "age = 2")) == 1
-
-    def test_create_table_single_column_is_primary_key(self):
-        """A table can be created with a single column, which becomes the primary key."""
-        from rhydb import Database
-
-        db = Database()
-        db.create_table(table_name="samples", columns=[{"name": "id", "type": "string"}])
-        db.append_data_from_string("samples", '{"id": "s1"}')
-
-        result = db.query("samples")
-        assert result.num_rows == 1
-        assert "id" in result.column_names
-
-    def test_create_table_empty_table_name_raises(self, empty_database):
-        with pytest.raises(ValueError, match="table_name cannot be empty"):
-            empty_database.create_table("", [{"name": "id", "type": "string"}])
-
-    def test_create_table_no_columns_raises(self, empty_database):
-        with pytest.raises(ValueError, match="at least one column"):
-            empty_database.create_table("samples", [])
-
-    def test_create_table_non_string_primary_key_raises(self, empty_database):
-        with pytest.raises(RuntimeError, match="must be of type 'string'"):
-            empty_database.create_table("samples", [{"name": "id", "type": "int"}])
-
-    def test_create_table_unknown_type_raises(self, empty_database):
-        with pytest.raises(ValueError, match="unknown type"):
-            empty_database.create_table("samples", [{"name": "x", "type": "nonsense"}])
-
-    def test_create_table_missing_name_raises(self, empty_database):
-        with pytest.raises(ValueError, match="non-empty string 'name'"):
-            empty_database.create_table("samples", [{"type": "int"}])
-
-    def test_create_table_column_not_dict_raises(self, empty_database):
-        with pytest.raises(TypeError, match="each column must be a dict"):
-            empty_database.create_table("samples", ["age"])
-
-    def test_create_table_duplicate_column_raises(self, empty_database):
-        with pytest.raises(RuntimeError, match="Duplicate column name"):
-            empty_database.create_table(
-                "samples", [{"name": "x", "type": "string"}, {"name": "x", "type": "int"}]
-            )
-
-
-class TestGetTables:
-    """Test the get_tables method."""
-
-    def test_get_tables_empty_database(self, empty_database):
-        """Even an empty database lists the built-in, always-present `reference_genomes` table."""
-        result = empty_database.get_tables()
-        assert isinstance(result, pa.Table)
-        assert "table_name" in result.column_names
-        assert set(result.to_pydict()["table_name"]) == {"reference_genomes"}
-
-    def test_get_tables_single_table(self, empty_database, main_reference_sequence):
-        """Test get_tables with one table."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-
-        result = empty_database.get_tables()
-        assert isinstance(result, pa.Table)
-        assert "table_name" in result.column_names
-        # The built-in `reference_genomes` table is a normal, visible table.
-        table_names = set(result.to_pydict()["table_name"])
-        assert table_names == {"sequences", "reference_genomes"}
-
-    def test_get_tables_multiple_tables(self, empty_database, main_reference_sequence):
-        """Test get_tables with multiple tables."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        create_gene_table(empty_database,
-            table_name="genes",
-            primary_key_name="id",
-            gene_name="S",
-            reference_sequence="MFVFLVLLPLVSSQCVNLTTRTQLPPAYTNSFTRGVYYPDKVFRSSVLHSTQDLFLPFFSNVTWFHAI*"
-        )
-
-        result = empty_database.get_tables()
-        assert isinstance(result, pa.Table)
-        # The built-in `reference_genomes` table is a normal, visible table.
-        table_names = set(result.to_pydict()["table_name"])
-        assert table_names == {"sequences", "genes", "reference_genomes"}
-
-
-class TestQuery:
-    """Test the query method that returns PyArrow Tables."""
-
-    def test_query_returns_pyarrow_table(self, empty_database, main_reference_sequence):
-        """Test that query returns a PyArrow Table."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        query = 'sequences'
-
-        result = empty_database.query(query)
-
-        assert isinstance(result, pa.Table)
-
-    def test_query_has_correct_schema(self, empty_database, main_reference_sequence):
-        """Test that the returned table has expected columns."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        query = 'sequences'
-        result = empty_database.query(query)
-
-        # Should have at least the primary key column
-        assert "primary_key" in result.column_names
-
-    def test_query_returns_data(self, empty_database, main_reference_sequence):
-        """Test that query returns rows."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        query = 'sequences'
-        result = empty_database.query(query)
-
-        assert result.num_rows > 0
-
-    def test_query_with_filter(self, empty_database, main_reference_sequence):
-        """Test query with a filter expression."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        # Get all rows first
-        all_query = 'sequences'
-        all_result = empty_database.query(all_query)
-
-        # Get filtered rows (False filter should return 0 rows)
-        filtered_query = 'sequences.filter(false)'
-        filtered_result = empty_database.query(filtered_query)
-
-        assert filtered_result.num_rows == 0
-        assert all_result.num_rows > filtered_result.num_rows
-
-    def test_query_to_batches(self, empty_database, main_reference_sequence):
-        """Test that the result can be converted to RecordBatches."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        query = 'sequences'
-        result = empty_database.query(query)
-
-        batches = result.to_batches()
-        assert isinstance(batches, list)
-        for batch in batches:
-            assert isinstance(batch, pa.RecordBatch)
-
-    def test_query_to_pydict(self, empty_database, main_reference_sequence):
-        """Test that the result can be converted to Python dict."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        query = 'sequences'
-        result = empty_database.query(query)
-
-        data = result.to_pydict()
-        assert isinstance(data, dict)
-        assert "primary_key" in data
-        assert isinstance(data["primary_key"], list)
-
-    def test_query_empty_query_raises(self, empty_database, main_reference_sequence):
-        """Test that empty query raises ValueError."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-
-        with pytest.raises(ValueError, match="query_string cannot be empty"):
-            empty_database.query("")
-
-    def test_query_invalid_query_raises(self, empty_database, main_reference_sequence):
-        """Test that invalid SaneQL raises an error."""
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-
-        with pytest.raises(RuntimeError):
-            empty_database.query("not valid saneql !")
-
-    def test_query_simple_database(self):
-        """Test query with a simple in-memory database."""
-        from rhydb import Database
-
-        db = Database()
-        create_nucleotide_sequence_table(db,
-            table_name="test",
-            primary_key_name="id",
-            sequence_name="seq",
-            reference_sequence="ACGT"
-        )
-        db.append_data_from_string("test", '{"id": "sample1", "seq": {"sequence": "AAAA", "insertions": []}}')
-        db.append_data_from_string("test", '{"id": "sample2", "seq": {"sequence": "CCCC", "insertions": []}}')
-
-        query = 'test'
-        result = db.query(query)
-
-        assert isinstance(result, pa.Table)
-        assert result.num_rows == 2
-        assert "id" in result.column_names
-
-        # Verify the data
-        data = result.to_pydict()
-        assert set(data["id"]) == {"sample1", "sample2"}
-
-    def test_query_preserves_data_after_checkpoint(self, empty_database, main_reference_sequence, temp_dir):
-        """Test that query works correctly after loading from checkpoint."""
-        from rhydb import Database
-
-        # Create and populate database
-        create_nucleotide_sequence_table(empty_database,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence
-        )
-        empty_database.append_data_from_file("sequences", INPUT_FILE)
-
-        # Query before checkpoint
-        query = 'sequences'
-        result_before = empty_database.query(query)
-
-        # Save and reload
-        save_path = os.path.join(temp_dir, "checkpoint")
-        empty_database.save_checkpoint(save_path)
-        loaded_db = Database(save_path)
-
-        # Query after checkpoint
-        result_after = loaded_db.query(query)
-
-        # Results should match
-        assert result_before.num_rows == result_after.num_rows
-        assert result_before.column_names == result_after.column_names
-
-
-class TestUpdateColumn:
-    """Test the update_column binding.
-
-    Extra string columns added alongside a sequence column are plain (non-indexed, non-phylo)
-    string columns, so these tests cover both the binding layer itself (argument validation,
-    marshalling, translation of C++ errors into Python exceptions) and the end-to-end update of a
-    string column. Scalar value column updates are additionally exercised via create_table in
-    TestCreateTable and against the loaded database in TestUpdateColumnOnLoadedDatabase.
-    """
-
-    def _database_with_string_column(self, main_reference_sequence):
-        from rhydb import Database
-
-        db = Database()
-        create_nucleotide_sequence_table(db,
-            table_name="sequences",
-            primary_key_name="primary_key",
-            sequence_name="main",
-            reference_sequence=main_reference_sequence,
-            extra_columns=["country"]
-        )
-        db.append_data_from_file("sequences", INPUT_FILE)
-        return db
-
-    @staticmethod
-    def _count(database, filter_expression):
-        return len(database.get_filtered_bitmap("sequences", filter_expression))
-
-    def test_update_column_empty_table_name_raises(self, empty_database):
-        """Test that an empty table name raises ValueError before reaching C++."""
-        with pytest.raises(ValueError, match="table_name cannot be empty"):
-            empty_database.update_column("", "country", "'x'")
-
-    def test_update_column_empty_column_name_raises(self, empty_database):
-        """Test that an empty column name raises ValueError before reaching C++."""
-        with pytest.raises(ValueError, match="column_name cannot be empty"):
-            empty_database.update_column("sequences", "", "'x'")
-
-    def test_update_column_unknown_column_raises(self, main_reference_sequence):
-        """Test that updating a non-existent column surfaces the C++ query error as ValueError."""
-        db = self._database_with_string_column(main_reference_sequence)
-        with pytest.raises(ValueError, match="does not contain a column"):
-            db.update_column("sequences", "does_not_exist", "1")
-
-    def test_update_string_column_all_rows(self, main_reference_sequence):
-        """A quoted string literal is assigned to every matched row of a string column."""
-        db = self._database_with_string_column(main_reference_sequence)
-        total = self._count(db, "true")
-        assert self._count(db, "country = 'Switzerland'") == total
-
-        db.update_column("sequences", "country", "'Germany'")
-
-        assert self._count(db, "country = 'Switzerland'") == 0
-        assert self._count(db, "country = 'Germany'") == total
-
-    def test_update_string_column_with_filter(self, main_reference_sequence):
-        """Only rows matching the filter are reassigned; a brand-new value is interned."""
-        db = self._database_with_string_column(main_reference_sequence)
-        total = self._count(db, "true")
-        assert total > 1
-
-        db.update_column("sequences", "country", "'France'", "primary_key = 'key_29'")
-
-        assert self._count(db, "country = 'France'") == 1
-        assert self._count(db, "country = 'Switzerland'") == total - 1
-
-    def test_update_string_column_clears_rows_to_null(self, main_reference_sequence):
-        """The literal 'null' clears the matched rows of a string column."""
-        db = self._database_with_string_column(main_reference_sequence)
-        assert self._count(db, "country = null") == 0
-
-        db.update_column("sequences", "country", "null", "primary_key = 'key_29'")
-
-        assert self._count(db, "country = null") == 1
-
-    def test_update_string_column_type_mismatch_raises(self, main_reference_sequence):
-        """A non-string literal is rejected for a string column."""
-        db = self._database_with_string_column(main_reference_sequence)
-        with pytest.raises(ValueError, match="expected string literal"):
-            db.update_column("sequences", "country", "5")
-
-
-class TestUpdateColumnOnLoadedDatabase:
-    """End-to-end update_column tests against the serialized database checked into the repo.
-
-    That database has real scalar value columns (age:int, qc_value:float, date:date,
-    test_boolean_column:bool) plus indexed string columns, and exercises the on-disk load path.
-    update_column mutates only the in-memory database (the on-disk state is never saved), so every
-    test loads its own fresh copy and the repo files are left untouched.
-    """
-
-    @pytest.fixture
-    def loaded_database(self):
-        from rhydb import Database
-        return Database(SERIALIZED_STATE_DIR)
-
-    @staticmethod
-    def _column(database, name):
-        return database.query(f'default.project({{{name}}})').to_pydict()[name]
-
-    @staticmethod
-    def _count(database, filter_expression):
-        return len(database.get_filtered_bitmap("default", filter_expression))
-
-    def test_update_int_column_all_rows(self, loaded_database):
-        """Updating without a filter assigns the value to every row."""
-        total = self._count(loaded_database, "true")
-        assert total > 0
-
-        loaded_database.update_column("default", "age", "42")
-
-        assert self._count(loaded_database, "age = 42") == total
-        assert all(age == 42 for age in self._column(loaded_database, "age"))
-
-    def test_update_int_column_with_filter(self, loaded_database):
-        """Only rows matching the filter are updated; the rest keep their values."""
-        matching = self._count(loaded_database, "age = 4")
-        total = self._count(loaded_database, "true")
-        assert 0 < matching < total
-
-        loaded_database.update_column("default", "age", "100", "age = 4")
-
-        assert self._count(loaded_database, "age = 4") == 0
-        assert self._count(loaded_database, "age = 100") == matching
-        # Rows that did not match still exist and were not turned into 100.
-        assert self._count(loaded_database, "age = 100") < total
-
-    def test_update_float_column(self, loaded_database):
-        """Float literals are assigned to a float column."""
-        loaded_database.update_column("default", "qc_value", "0.5")
-        assert all(abs(value - 0.5) < 1e-9 for value in self._column(loaded_database, "qc_value"))
-
-    def test_update_bool_column(self, loaded_database):
-        """Boolean literals are assigned to a bool column (including rows that were null)."""
-        loaded_database.update_column("default", "test_boolean_column", "false")
-        assert all(value is False for value in self._column(loaded_database, "test_boolean_column"))
-
-    def test_update_date_column(self, loaded_database):
-        """SaneQL date literals are assigned to a date column."""
-        loaded_database.update_column("default", "date", "'2000-01-01'::date")
-        assert all(
-            value == datetime.date(2000, 1, 1) for value in self._column(loaded_database, "date")
-        )
-
-    def test_update_column_clears_rows_to_null(self, loaded_database):
-        """The literal 'null' clears the matched rows."""
-        matching = self._count(loaded_database, "age = 4")
-        nulls_before = self._count(loaded_database, "age = null")
-        assert matching > 0
-
-        loaded_database.update_column("default", "age", "null", "age = 4")
-
-        assert self._count(loaded_database, "age = 4") == 0
-        assert self._count(loaded_database, "age = null") == nulls_before + matching
-
-    def test_update_does_not_persist_to_disk(self, loaded_database):
-        """Updating the in-memory database must not modify the on-disk repo state."""
-        loaded_database.update_column("default", "age", "999")
-        assert self._count(loaded_database, "age = 999") > 0
-
-        from rhydb import Database
-        fresh = Database(SERIALIZED_STATE_DIR)
-        assert self._count(fresh, "age = 999") == 0
